Refactoring
- Moved into controller/service/repository package
- Removed R class
