+++ user-service/src/main/kotlin/com/hulkdx/findprofessional/controller/AuthController.kt
@@ -1,17 +1,28 @@
-package com.hulkdx.findprofessional
+package com.hulkdx.findprofessional.controller
 
-import com.hulkdx.findprofessional.model.request.RegisterRequest
-import com.hulkdx.findprofessional.model.response.AuthResponse
+import com.hulkdx.findprofessional.model.ApiError
 import com.hulkdx.findprofessional.model.request.LoginRequest
 import com.hulkdx.findprofessional.model.request.RefreshRequest
-import com.hulkdx.findprofessional.utils.toNormalUserResponse
-import com.hulkdx.findprofessional.utils.R
+import com.hulkdx.findprofessional.model.request.RegisterRequest
+import com.hulkdx.findprofessional.model.response.AuthResponse
+import com.hulkdx.findprofessional.service.AuthService
+import com.hulkdx.findprofessional.service.RefreshService
+import com.hulkdx.findprofessional.service.TokenService
+import com.hulkdx.findprofessional.utils.Errors.EMAIL_EXISTS
+import com.hulkdx.findprofessional.utils.Errors.EMAIL_NOT_VALID
+import com.hulkdx.findprofessional.utils.Errors.INVALID_TOKEN_TYPE
+import com.hulkdx.findprofessional.utils.Errors.PASSWORD_NOT_VALID
 import com.hulkdx.findprofessional.utils.Validator
+import com.hulkdx.findprofessional.utils.toNormalUserResponse
 import com.hulkdx.findprofessional.utils.toUserResponse
 import jakarta.validation.Valid
 import org.springframework.dao.DataIntegrityViolationException
 import org.springframework.data.r2dbc.config.EnableR2dbcAuditing
 import org.springframework.http.HttpHeaders
+import org.springframework.http.HttpStatus.BAD_REQUEST
+import org.springframework.http.HttpStatus.CONFLICT
+import org.springframework.http.HttpStatus.OK
+import org.springframework.http.HttpStatus.UNAUTHORIZED
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
@@ -32,39 +43,41 @@ class AuthController(
     private val tokenService: TokenService,
     private val refreshService: RefreshService,
 ) {
-    private val emailNotValid = "Email is not valid"
-    private val passwordNotValid = "Password is not valid"
-    private val emailExists = "Email already exists"
-    private val invalidTokenType = "Invalid token type"
 
     @PostMapping("/register")
     suspend fun register(@RequestBody @Valid body: RegisterRequest): ResponseEntity<*> {
         if (!Validator.isEmailValid(body.email)) {
-            return R.badRequest(emailNotValid)
+            return ResponseEntity.status(BAD_REQUEST)
+                .body(ApiError(EMAIL_NOT_VALID))
         }
         if (!Validator.isPasswordValid(body.password)) {
-            return R.badRequest(passwordNotValid)
+            return ResponseEntity.status(BAD_REQUEST)
+                .body(ApiError(PASSWORD_NOT_VALID))
         }
         return try {
             val user = authService.register(body)
             val token = tokenService.createToken(user)
-            R.ok(body = AuthResponse(token, user.toNormalUserResponse()))
+            ResponseEntity.status(OK)
+                .body(AuthResponse(token, user.toNormalUserResponse()))
         } catch (e: DataIntegrityViolationException) {
-            R.conflict(emailExists)
+            ResponseEntity.status(CONFLICT)
+                .body(ApiError(EMAIL_EXISTS))
         }
     }
 
     @PostMapping("/login")
     suspend fun login(@RequestBody @Valid body: LoginRequest): ResponseEntity<*> {
         if (!Validator.isEmailValid(body.email)) {
-            return R.badRequest(emailNotValid)
+            return ResponseEntity.status(BAD_REQUEST)
+                .body(ApiError(EMAIL_NOT_VALID))
         }
         val user = authService.login(body)
         return if (user != null) {
             val token = tokenService.createToken(user)
-            R.ok(body = AuthResponse(token, user.toUserResponse()))
+            ResponseEntity.status(OK)
+                .body(AuthResponse(token, user.toUserResponse()))
         } else {
-            R.unauthorized()
+            ResponseEntity.status(UNAUTHORIZED).build<Unit>()
         }
     }
 
@@ -73,20 +86,15 @@ class AuthController(
         @RequestHeader(HttpHeaders.AUTHORIZATION) auth: String,
         @RequestBody @Valid request: RefreshRequest,
     ): ResponseEntity<*> {
-        val authSplit = auth.split(" ")
-        if (authSplit.size != 2) {
-            return R.badRequest(invalidTokenType)
-        }
-        val (authType, accessToken) = authSplit
-        if (authType != "Bearer") {
-            return R.badRequest(invalidTokenType)
-        }
+        val accessToken = authService.getAccessToken(auth)
+            ?: return ResponseEntity.status(BAD_REQUEST).body(ApiError(INVALID_TOKEN_TYPE))
 
         val body = refreshService.refreshToken(accessToken, request.refreshToken)
         return if (body == null) {
-            R.unauthorized()
+            ResponseEntity.status(UNAUTHORIZED).build<Unit>()
         } else {
-            R.ok(body)
+            ResponseEntity.status(OK)
+                .body(body)
         }
     }
 }
