+++ user-service/src/main/kotlin/com/hulkdx/findprofessional/repository/ProRepository.kt
@@ -1,7 +1,6 @@
-package com.hulkdx.findprofessional
+package com.hulkdx.findprofessional.repository
 
 import com.hulkdx.findprofessional.model.Professional
-import com.hulkdx.findprofessional.model.User
 import org.springframework.data.repository.kotlin.CoroutineCrudRepository
 
 interface ProRepository : CoroutineCrudRepository<Professional, Int> {
