+++ user-service/src/main/kotlin/com/hulkdx/findprofessional/repository/UserRepository.kt
@@ -1,4 +1,4 @@
-package com.hulkdx.findprofessional
+package com.hulkdx.findprofessional.repository
 
 import com.hulkdx.findprofessional.model.User
 import org.springframework.data.repository.kotlin.CoroutineCrudRepository
