+++ user-service/src/main/kotlin/com/hulkdx/findprofessional/utils/Errors.kt
@@ -1,0 +1,8 @@
+package com.hulkdx.findprofessional.utils
+
+object Errors {
+    const val EMAIL_NOT_VALID = "Email is not valid"
+    const val PASSWORD_NOT_VALID = "Password is not valid"
+    const val EMAIL_EXISTS = "Email already exists"
+    const val INVALID_TOKEN_TYPE = "Invalid token type"
+}
