+++ user-service/src/test/kotlin/com/hulkdx/findprofessional/RegisterTests.kt
@@ -1,9 +1,13 @@
 package com.hulkdx.findprofessional
 
+import com.hulkdx.findprofessional.controller.AuthController
 import com.hulkdx.findprofessional.model.response.AuthResponse
 import com.hulkdx.findprofessional.model.response.UserResponse
 import com.hulkdx.findprofessional.model.response.TokenResponse
 import com.hulkdx.findprofessional.model.User
+import com.hulkdx.findprofessional.repository.UserRepository
+import com.hulkdx.findprofessional.service.AuthService
+import com.hulkdx.findprofessional.service.TokenService
 import com.hulkdx.findprofessional.utils.TestPasswordEncoder
 import com.hulkdx.findprofessional.utils.createRegisterRequest
 import com.hulkdx.findprofessional.utils.errorMessage
