+++ user-service/src/test/kotlin/com/hulkdx/findprofessional/TokenServiceTest.kt
@@ -1,5 +1,6 @@
 package com.hulkdx.findprofessional
 
+import com.hulkdx.findprofessional.service.TokenService
 import com.hulkdx.findprofessional.utils.createJwt
 import kotlinx.coroutines.reactor.mono
 import kotlinx.coroutines.test.runTest
